used users dropdown component in the add voter modal
